テストユーザーでアカウント編集できないように権限を追加

+++ app/javascript/packs/components/pages/account/edit.vue
@@ -5,40 +5,42 @@
 
       <div class="d-lg-block d-md-block d-sm-none d-none">
         <my-dashboard :menus="menus" @pageChange="pageChange">
-          <component :is="menuComponent" :parentErrors="errors" @emailChange="pageChange('my-email-edit')" @update="update" @destroy="destroy" :isLoading="isLoading"></component>
+          <component
+            :is="menuComponent"
+            :parentErrors="errors"
+            @emailChange="pageChange('my-email-edit')"
+            @update="update"
+            @destroy="destroy"
+            :isLoading="isLoading"
+          ></component>
         </my-dashboard>
       </div>
 
       <div class="d-lg-none d-md-none d-sm-block d-block">
         <my-bottom-btn @bottomBtn="isActive = !isActive">mdi-format-list-bulleted</my-bottom-btn>
-          <v-navigation-drawer
-            v-model="isActive"
-            color="#1FAB89"
-            fixed
-            right
-            dark
-          >
-
+        <v-navigation-drawer v-model="isActive" color="#1FAB89" fixed right dark>
           <template>
             <v-list-item>
               <v-list-item-icon>
                 <v-icon>mdi-cogs</v-icon>
               </v-list-item-icon>
-              <v-list-item-title><strong>設定</strong></v-list-item-title>
+              <v-list-item-title>
+                <strong>設定</strong>
+              </v-list-item-title>
             </v-list-item>
           </template>
 
-            <v-list>
-              <v-list-item-group v-model="group">
-                <v-list-item v-for="menu in menus" :key="menu.page" @click="pageChange(menu.page)">
-                  <v-list-item-icon>
-                    <v-icon>{{ menu.icon }}</v-icon>
-                  </v-list-item-icon>
-                  <v-list-item-content>{{ menu.name }}</v-list-item-content>
-                </v-list-item>
-              </v-list-item-group>
-            </v-list>
-          </v-navigation-drawer>
+          <v-list>
+            <v-list-item-group v-model="group">
+              <v-list-item v-for="menu in menus" :key="menu.page" @click="pageChange(menu.page)">
+                <v-list-item-icon>
+                  <v-icon>{{ menu.icon }}</v-icon>
+                </v-list-item-icon>
+                <v-list-item-content>{{ menu.name }}</v-list-item-content>
+              </v-list-item>
+            </v-list-item-group>
+          </v-list>
+        </v-navigation-drawer>
         <v-row justify="center">
           <v-col sm="10" xs="12">
             <component :is="menuComponent" @emailChange="pageChange('my-email-edit')"></component>
@@ -61,12 +63,15 @@ import { createNamespacedHelpers } from "vuex";
 import { mapActions } from "vuex";
 
 const {
-  mapActions: mapActionsOfAccount
+  mapActions: mapActionsOfAccount,
+  mapState: mapStateOfAccount
 } = createNamespacedHelpers("Account");
 
-const {
-  mapMutations: mapMutationsOfSession
-} = createNamespacedHelpers("Session");
+const { mapMutations: mapMutationsOfSession } = createNamespacedHelpers(
+  "Session"
+);
+
+const { mapActions: mapActionsOfAlert } = createNamespacedHelpers("Alert");
 
 export default {
   components: {
@@ -86,31 +91,58 @@ export default {
       errors: {},
       menus: [
         { name: "プロフィール", icon: "mdi-account", page: "my-profile" },
-        { name: "パスワード", icon: "mdi-shield-account", page: "my-password-edit" },
-        { name: "サービス退会", icon: "mdi-account-remove", page: "my-service-end" }
+        {
+          name: "パスワード",
+          icon: "mdi-shield-account",
+          page: "my-password-edit"
+        },
+        {
+          name: "サービス退会",
+          icon: "mdi-account-remove",
+          page: "my-service-end"
+        }
       ],
       menuComponent: "my-profile"
-    }
+    };
+  },
+  computed: {
+    ...mapStateOfAccount(["user"])
   },
   methods: {
     ...mapActionsOfAccount(["updateUser", "destroyUser"]),
     ...mapMutationsOfSession(["setId"]),
     ...mapActions(["reset"]),
+    ...mapActionsOfAlert(["setAlert"]),
+
     pageChange(page) {
       this.menuComponent = page;
     },
     async update(e) {
+      if (this.user.authority == 1) {
+        this.setAlert({
+          msg: "テストユーザーではアカウントを編集することができません。",
+          type: "warning"
+        });
+        return;
+      }
       this.isLoading = true;
       let data = await this.updateUser(e);
       this.isLoading = false;
 
       if (data.result == 1) {
         for (let key in data.errors) {
-        this.$set(this.errors, key, data.errors[key]);
-      }
+          this.$set(this.errors, key, data.errors[key]);
+        }
       }
     },
     async destroy() {
+      if (this.user.authority == 1) {
+        this.setAlert({
+          msg: "テストユーザーではアカウントを削除することができません。",
+          type: "warning"
+        });
+        return;
+      }
       this.isLoading = true;
       let result = await this.destroyUser();
       this.isLoading = false;
@@ -130,7 +162,7 @@ export default {
       }
     }
   }
-}
+};
 </script>
 
 <style>
